// fem-vision-studio/design-mockup/src/lib/types/steel/VerloopRapport.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MaatgevendPunt } from "./MaatgevendPunt";
import type { StabiliteitsDoorsnede } from "./StabiliteitsDoorsnede";
import type { Toetsdoorsnede } from "./Toetsdoorsnede";
import type { VerloopMaten } from "./VerloopMaten";

/**
 * Alles wat het rapport over een verlopende staaf moet tonen.
 */
export type VerloopRapport = { begin_naam: string, eind_naam: string, begin: VerloopMaten, eind: VerloopMaten, 
/**
 * Aantal rekenpunten (unieke posities) waarop de doorsnedetoetsen zijn uitgevoerd.
 */
aantal_rekenpunten: number, 
/**
 * De zes toetsdoorsneden op x ≈ 0, L/5, …, L (het dichtstbijzijnde rekenpunt).
 */
toetsdoorsneden: Array<Toetsdoorsnede>, maatgevend?: MaatgevendPunt, stabiliteit: Array<StabiliteitsDoorsnede>, 
/**
 * Toelichtingen bij het verloop als geheel (gelaste aanname, f_y-dikte, …).
 */
notities: Array<string>, };